backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-multipart>=0.0.9
openai>=1.60.0,<2.0.0
tiktoken>=0.8.0,<1.0.0
faiss-cpu>=1.9.0,<2.0.0
rank-bm25>=0.2.2,<1.0.0
sentence-transformers>=3.4.0,<4.0.0
PyMuPDF>=1.25.0,<2.0.0
python-docx>=1.1.0,<2.0.0
markdown>=3.7,<4.0.0
beautifulsoup4>=4.12.0,<5.0.0
pydantic-settings>=2.7.0,<3.0.0
aiofiles>=24.0.0
